+++ yeat/cli/cli.py
@@ -13,20 +13,36 @@ from argparse import Action, ArgumentParser
 import json
 import sys
 import yeat
-from yeat.bandage import bandage
-from yeat.workflows import workflows
+from yeat import workflows
 
 
-CONFIG_TEMPLATE = [
-    dict(
-        algorithm="spades",
-        extra_args="--meta",
-    ),
-    dict(
-        algorithm="megahit",
-        extra_args="--min-count 5 --min-contig-len 300",
-    ),
-]
+CONFIG_TEMPLATE = {
+    "samples": {
+        "sample1": ["path_to_read1", "path_to_read2"],
+        "sample2": ["path_to_read1", "path_to_read2"],
+        "sample3": ["path_to_read"],
+    },
+    "assemblers": [
+        {
+            "label": "spades-meta",
+            "algorithm": "spades",
+            "extra_args": "--meta",
+            "samples": ["sample1"],
+        },
+        {
+            "label": "megahit-mins",
+            "algorithm": "megahit",
+            "extra_args": "--min-count 5 --min-contig-len 300",
+            "samples": ["sample1", "sample2"],
+        },
+        {
+            "label": "pacbio-hifi",
+            "algorithm": "canu",
+            "extra_args": "genomeSize=4.8m",
+            "samples": ["sample3"],
+        },
+    ],
+}
 
 
 class InitAction(Action):
@@ -37,17 +53,12 @@ class InitAction(Action):
 
 
 def options(parser):
+    parser.add_argument("-v", "--version", action="version", version=f"YEAT v{yeat.__version__}")
     parser.add_argument(
         "--init",
         action=InitAction,
         nargs=0,
         help="print a template assembly config file to the terminal (stdout) and exit",
-    )
-    parser.add_argument(
-        "-n",
-        "--dry-run",
-        action="store_true",
-        help="construct workflow DAG and print a summary but do not execute",
     )
     parser.add_argument(
         "-o",
@@ -58,13 +69,6 @@ def options(parser):
         help="output directory; default is current working directory",
     )
     parser.add_argument(
-        "--sample",
-        type=str,
-        metavar="S",
-        default="sample",
-        help="specify a unique sample name S for storing assembly results in the working directory; by default S=sample",
-    )
-    parser.add_argument(
         "-t",
         "--threads",
         type=int,
@@ -72,37 +76,24 @@ def options(parser):
         default=1,
         help="execute workflow with T threads; by default T=1",
     )
-    parser.add_argument("-v", "--version", action="version", version=f"YEAT v{yeat.__version__}")
-
-
-def read_input_types(parser):
-    read_input_options = parser.add_argument_group("input read options")
-    mx = read_input_options.add_mutually_exclusive_group(required=True)
-    mx.add_argument(
-        "--paired",
-        metavar=("READ1", "READ2"),
-        type=str,
-        nargs=2,
-        help="paired-end reads in FASTQ format",
+    parser.add_argument(
+        "-n",
+        "--dry-run",
+        action="store_true",
+        help="construct workflow DAG and print a summary but do not execute",
     )
-    mx.add_argument("--pacbio", metavar="READ", type=str, help="PacBio HiFi-reads in FASTQ format")
 
 
 def get_parser(exit_on_error=True):
     parser = ArgumentParser(exit_on_error=exit_on_error)
     options(parser)
     downsample.options(parser)
-    read_input_types(parser)
     parser.add_argument("config", type=str, help="config file")
     return parser
 
 
 def main(args=None):
     if args is None:
-        args = get_parser().parse_args()
-    assembly_configs = AssemblerConfig.parse_json(open(args.config))
-    workflows.run_workflow(args, assembly_configs)
-    if not args.dry_run:
-        bandage.run_bandage(
-            assembly_configs=assembly_configs, outdir=args.outdir, cores=args.threads
-        )
+        args = get_parser().parse_args()  # pragma: no cover
+    config = AssemblerConfig.from_json(args.config, args.threads)
+    workflows.run_workflows(args, config)