+++ yeat/tests/test_bandage.py
@@ -14,13 +14,20 @@ import pytest
 import shutil
 import subprocess
 from unittest.mock import patch
-from yeat.bandage import bandage
 from yeat.tests import data_file
+from yeat.workflows import bandage
 
 
 @pytest.mark.bandage
 def test_check_bandage():
     assert bandage.check_bandage()
+
+
+@pytest.mark.bandage
+@patch("subprocess.run")
+def test_bandage_subprocess_failed(function_mock):
+    function_mock.return_value = subprocess.CompletedProcess(["Bandage", "--help"], 1)
+    assert bandage.check_bandage() == False
 
 
 @patch.dict(os.environ, {"PATH": "ROUGE"})
@@ -30,11 +37,11 @@ def test_env_path_has_no_path_to_bandage(capsys):
     assert "No such file or directory: 'Bandage'" in out
 
 
-@patch("yeat.bandage.bandage.check_bandage")
+@patch("yeat.workflows.bandage.check_bandage")
 def test_no_bandage_warning(function_mock):
     function_mock.return_value = False
     with pytest.warns(UserWarning, match=r"Unable to run Bandage; skipping Bandage"):
-        bandage.run_bandage(assembly_configs=[])
+        bandage.run_bandage(None, None)
 
 
 @pytest.mark.bandage