+++ yeat/tests/test_cli.py
@@ -11,7 +11,7 @@ from argparse import ArgumentError
 import json
 import pytest
 from yeat import cli
-from yeat.cli import InitAction
+from yeat.cli import InitAction, downsample
 from yeat.tests import data_file
 
 
@@ -27,9 +27,6 @@ def test_invalid_custom_coverage_negative(coverage):
     arglist = [
         "--coverage",
         coverage,
-        "--paired",
-        data_file("short_reads_1.fastq.gz"),
-        data_file("short_reads_2.fastq.gz"),
         data_file("megahit.cfg"),
     ]
     with pytest.raises(ArgumentError, match=rf"{coverage} is not a positive integer"):
@@ -41,10 +38,15 @@ def test_invalid_custom_coverage_noninteger(coverage):
     arglist = [
         "--coverage",
         coverage,
-        "--paired",
-        data_file("short_reads_1.fastq.gz"),
-        data_file("short_reads_2.fastq.gz"),
         data_file("megahit.cfg"),
     ]
     with pytest.raises(ArgumentError, match=rf"{coverage} is not an integer"):
         args = cli.get_parser(exit_on_error=False).parse_args(arglist)
+
+
+@pytest.mark.parametrize(
+    "value",
+    [1, 10, 100],
+)
+def test_check_positive(value):
+    downsample.check_positive(value) == value