+++ yeat/tests/test_pacbio.py
@@ -7,14 +7,13 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
-import multiprocessing
 from pathlib import Path
 import pytest
-from yeat import cli, workflows
-from yeat.tests import data_file
+from yeat import cli
+from yeat.tests import data_file, get_core_count
 
 
-def test_pacbio_hifi_read_assemblers_dry_run(tmp_path):
+def test_pacbio_hifi_assemblers_dry_run(tmp_path):
     wd = str(tmp_path)
     arglist = [
         "--outdir",
@@ -22,58 +21,31 @@ def test_pacbio_hifi_read_assemblers_dry_run(tmp_path):
         "-n",
         "-t",
         "4",
-        "--pacbio",
-        data_file("ecoli.fastq.gz"),
-        data_file("pacbio.cfg"),
+        data_file("configs/pacbio.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
 
 
-def test_invalid_read_file():
-    with pytest.raises(Exception, match=r"No such file:.*read\'"):
-        read = "read"
-        assemblers = ["flye"]
-        workflows.run_pacbio(read, assemblers)
-
-
-@pytest.mark.parametrize(
-    "extra_args,cores,expected",
-    [
-        ("", 4, r"Missing required input argument from config: 'genomeSize'"),
-        (
-            "genomeSize=4.8m",
-            1,
-            r"Canu requires at least 4 avaliable cores; increase `--threads` to 4 or more",
-        ),
-    ],
-)
-def test_check_canu_required_params_errors(extra_args, cores, expected):
-    with pytest.raises(ValueError, match=expected):
-        workflows.check_canu_required_params(extra_args, cores)
-
-
 @pytest.mark.hifi
 @pytest.mark.parametrize(
-    "algorithm,finalcontigs",
+    "algorithm,label,expected",
     [
-        ("canu", "sample.contigs.fasta"),
-        ("flye", "assembly.fasta"),
+        ("canu", "canu-default", "sample1.contigs.fasta"),
+        ("flye", "flye-default", "assembly.fasta"),
     ],
 )
-def test_pacbio_hifi_read_assemblers(algorithm, finalcontigs, capsys, tmp_path):
+def test_pacbio_hifi_read_assemblers(algorithm, label, expected, capsys, tmp_path):
     wd = str(tmp_path)
-    cores = str(multiprocessing.cpu_count())
+    cores = str(get_core_count())
     arglist = [
         "--outdir",
         wd,
         "--threads",
         cores,
-        "--pacbio",
-        data_file("ecoli.fastq.gz"),
-        data_file(f"{algorithm}.cfg"),
+        data_file(f"configs/{algorithm}.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
-    result = Path(wd).resolve() / "analysis" / algorithm / finalcontigs
-    assert result.exists()
+    observed = Path(wd).resolve() / "analysis" / "sample1" / label / algorithm / expected
+    assert observed.exists()