+++ yeat/tests/test_paired.py
@@ -7,83 +7,81 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
+import json
 import pandas as pd
 from pathlib import Path
 import pytest
 import re
 import subprocess
-from yeat import cli, workflows
+from yeat import cli
 from yeat.tests import data_file
 
 
-def test_paired_end_read_assemblers_dry_run(tmp_path):
+def test_paired_end_assemblers_dry_run(tmp_path):
     """The purpose of this test is to check if the snakemake workflow is properly defined."""
     wd = str(tmp_path)
     arglist = [
         "--outdir",
         wd,
-        "--sample",
-        "Animal_289",
         "-n",
-        "--paired",
-        data_file("Animal_289_R1.fq.gz"),
-        data_file("Animal_289_R2.fq.gz"),
-        data_file("config.cfg"),
+        data_file("configs/paired.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
 
 
-def test_invalid_read1_file():
-    with pytest.raises(Exception, match=r"No such file:.*read1\'"):
-        read1 = "read1"
-        read2 = "read2"
-        assemblers = ["spades"]
-        workflows.run_paired(read1, read2, assemblers)
-
-
-def test_invalid_read2_file():
-    with pytest.raises(Exception, match=r"No such file:.*read2\'"):
-        read1 = data_file("short_reads_1.fastq.gz")
-        read2 = "read2"
-        assemblers = ["spades"]
-        workflows.run_paired(read1, read2, assemblers)
-
-
 @pytest.mark.long
-def test_multiple_assemblers(capsys, tmp_path):
+@pytest.mark.parametrize(
+    "algorithm,label,expected",
+    [
+        ("spades", "spades-meta", "contigs.fasta"),
+        ("megahit", "megahit-default", "final.contigs.fa"),
+        ("unicycler", "unicycler-default", "assembly.fasta"),
+    ],
+)
+def test_individual_paired_end_assemblers(algorithm, label, expected, capsys, tmp_path):
     wd = str(tmp_path)
     arglist = [
         "--outdir",
         wd,
-        "--paired",
-        data_file("short_reads_1.fastq.gz"),
-        data_file("short_reads_2.fastq.gz"),
-        data_file("config.cfg"),
+        data_file(f"configs/{algorithm}.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
-    spades_result = Path(wd).resolve() / "analysis" / "spades" / "contigs.fasta"
-    assert spades_result.exists()
-    megahit_result = Path(wd).resolve() / "analysis" / "megahit" / "final.contigs.fa"
-    assert megahit_result.exists()
+    analysis_dir = Path(wd).resolve() / "analysis"
+    assert (analysis_dir / "sample1" / label / algorithm / expected).exists()
 
 
 @pytest.mark.long
-def test_unicycler(capsys, tmp_path):
+def test_multiple_spades(capsys, tmp_path):
     wd = str(tmp_path)
     arglist = [
         "--outdir",
         wd,
-        "--paired",
-        data_file("short_reads_1.fastq.gz"),
-        data_file("short_reads_2.fastq.gz"),
-        data_file("unicycler.cfg"),
+        data_file(f"configs/two_spades.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
-    assembly_result = Path(wd).resolve() / "analysis" / "unicycler" / "assembly.fasta"
-    assert assembly_result.exists()
+    analysis_dir = Path(wd).resolve() / "analysis"
+    assert (analysis_dir / "sample1" / "spades-default" / "spades" / "contigs.fasta").exists()
+    assert (analysis_dir / "sample1" / "spades-meta" / "spades" / "contigs.fasta").exists()
+
+
+@pytest.mark.long
+def test_multiple_paired_end_assemblers(capsys, tmp_path):
+    wd = str(tmp_path)
+    arglist = [
+        "--outdir",
+        wd,
+        data_file(f"configs/unicycler_spades.cfg"),
+    ]
+    args = cli.get_parser().parse_args(arglist)
+    cli.main(args)
+    analysis_dir = Path(wd).resolve() / "analysis"
+    assert (
+        analysis_dir / "sample1" / "unicycler-default" / "unicycler" / "assembly.fasta"
+    ).exists()
+    assert (analysis_dir / "sample1" / "spades-meta" / "spades" / "contigs.fasta").exists()
 
 
 @pytest.mark.long
@@ -102,18 +100,18 @@ def test_custom_downsample_input(
         downsample,
         "--seed",
         "0",
-        "--paired",
-        data_file("short_reads_1.fastq.gz"),
-        data_file("short_reads_2.fastq.gz"),
-        data_file("megahit.cfg"),
+        data_file("configs/megahit.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
-    quast_report = Path(wd).resolve() / "analysis" / "quast" / "megahit" / "report.tsv"
+    analysis_dir = Path(wd).resolve() / "analysis"
+    quast_report = (
+        analysis_dir / "sample1" / "megahit-default" / "megahit" / "quast" / "report.tsv"
+    )
     df = pd.read_csv(quast_report, sep="\t")
-    assert df.iloc[12]["sample_contigs"] == num_contigs
-    assert df.iloc[13]["sample_contigs"] == largest_contig
-    assert df.iloc[14]["sample_contigs"] == total_len
+    assert df.iloc[12]["sample1_contigs"] == num_contigs
+    assert df.iloc[13]["sample1_contigs"] == largest_contig
+    assert df.iloc[14]["sample1_contigs"] == total_len
 
 
 @pytest.mark.long
@@ -125,20 +123,20 @@ def test_custom_coverage_input(coverage, capsys, tmp_path):
         wd,
         "-c",
         coverage,
-        "--paired",
-        data_file("short_reads_1.fastq.gz"),
-        data_file("short_reads_2.fastq.gz"),
-        data_file("megahit.cfg"),
+        data_file("configs/megahit.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
-    quast_report = Path(wd).resolve() / "analysis" / "quast" / "megahit" / "report.tsv"
+    analysis_dir = Path(wd).resolve() / "analysis"
+    quast_report = (
+        analysis_dir / "sample1" / "megahit-default" / "megahit" / "quast" / "report.tsv"
+    )
     df = pd.read_csv(quast_report, sep="\t")
-    num_contigs = df.iloc[12]["sample_contigs"]
+    num_contigs = df.iloc[12]["sample1_contigs"]
     assert num_contigs == 56
-    largest_contig = df.iloc[13]["sample_contigs"]
+    largest_contig = df.iloc[13]["sample1_contigs"]
     assert largest_contig == 35168
-    total_len = df.iloc[14]["sample_contigs"]
+    total_len = df.iloc[14]["sample1_contigs"]
     assert total_len == 199940
 
 
@@ -151,20 +149,20 @@ def test_random_downsample_seed(execution_number, capsys, tmp_path):
         wd,
         "-d",
         "2000",
-        "--paired",
-        data_file("short_reads_1.fastq.gz"),
-        data_file("short_reads_2.fastq.gz"),
-        data_file("megahit.cfg"),
+        data_file("configs/megahit.cfg"),
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
-    quast_report = Path(wd).resolve() / "analysis" / "quast" / "megahit" / "report.tsv"
+    analysis_dir = Path(wd).resolve() / "analysis"
+    quast_report = (
+        analysis_dir / "sample1" / "megahit-default" / "megahit" / "quast" / "report.tsv"
+    )
     df = pd.read_csv(quast_report, sep="\t")
-    num_contigs = df.iloc[12]["sample_contigs"]
+    num_contigs = df.iloc[12]["sample1_contigs"]
     assert num_contigs == pytest.approx(76, abs=15)  # 76 +/- 20%
-    largest_contig = df.iloc[13]["sample_contigs"]
+    largest_contig = df.iloc[13]["sample1_contigs"]
     assert largest_contig == pytest.approx(5228, abs=1045)  # 5228 +/- 20%
-    total_len = df.iloc[14]["sample_contigs"]
+    total_len = df.iloc[14]["sample1_contigs"]
     assert total_len == pytest.approx(74393, abs=14878)  # 74393 +/- 20%
 
 
@@ -189,24 +187,25 @@ def prep_uncompressed_reads(filename, tmp_path):
     ],
 )
 def test_uncompressed_input_reads(inread1, inread2, capfd, tmp_path):
-    wd = tmp_path / "wd"
+    wd = str(tmp_path)
     inread1 = prep_uncompressed_reads(inread1, tmp_path)
     inread2 = prep_uncompressed_reads(inread2, tmp_path)
+    cfg_data = json.load(open(data_file("configs/megahit.cfg")))
+    cfg_data["samples"]["sample1"] = [inread1, inread2]
+    cfg = str(Path(wd).resolve() / "megahit.cfg")
+    json.dump(cfg_data, open(cfg, "w"))
     arglist = [
         "--outdir",
-        str(wd),
-        "--paired",
-        inread1,
-        inread2,
-        data_file("megahit.cfg"),
+        wd,
+        cfg,
     ]
     args = cli.get_parser().parse_args(arglist)
     cli.main(args)
-    outread1 = wd / "seq" / "input" / "sample_R1.fq.gz"
-    outread2 = wd / "seq" / "input" / "sample_R2.fq.gz"
+    outread1 = Path(wd) / "seq" / "input" / "sample1" / "sample1_R1.fq.gz"
+    outread2 = Path(wd) / "seq" / "input" / "sample1" / "sample1_R2.fq.gz"
     assert outread1.exists()
     assert outread2.exists()
     subprocess.run(["gzip", "-tv", outread1, outread2])
     captured = capfd.readouterr()
-    assert re.search(r"seq/input/sample_R1.fq.gz:\s*OK", captured.err)
-    assert re.search(r"seq/input/sample_R2.fq.gz:\s*OK", captured.err)
+    assert re.search(r"seq/input/sample1/sample1_R1.fq.gz:\s*OK", captured.err)
+    assert re.search(r"seq/input/sample1/sample1_R2.fq.gz:\s*OK", captured.err)